Update index.js & index.html
Added a timeout function so that after 3 seconds the modal is displayed on the page by changing its display properly from none to inline.
Added a function to close the modal when the close button is pressed.
Removed disabled from modal-close-btn in index.html, as this was preventing the modal being closed and came with the starter code provided.